// packages/core/src/payload-types.ts
export interface User {
  id: string;
  username?: string | null;
  avatarUrl?: string | null;
  onboarded?: boolean | null;
  role?: ('admin' | 'user' | 'demo')[] | null;
  tenants?:
    | {
        tenant: string | Tenant;
        role: string | Role;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
  enableAPIKey?: boolean | null;
  apiKey?: string | null;
  apiKeyIndex?: string | null;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tenants".
 */
export interface Tenant {
  id: string;
  name: string;
  slug: string;
  subdomain: string;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "roles".
 */
export interface Role {
  id: string;
  tenant?: (string | null) | Tenant;
  /**
   * Enter the name of the role.
   */
  name: string;
  /**
   * Enter description for the role.
   */
  description?: string | null;
  projects: {
    create: boolean;
    /**
     * Set to 0 for unlimited servers. This is only applicable if Create permission is enabled.
     */
    createLimit?: number | null;
    update: boolean;
    read: boolean;
    readLimit?: ('all' | 'createdByUser') | null;
    delete: boolean;
  };
  services: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  servers: {
    create: boolean;
    /**
     * Set to 0 for unlimited servers. This is only applicable if Create permission is enabled.
     */
    createLimit?: number | null;
    update: boolean;
    read: boolean;
    readLimit?: ('all' | 'createdByUser') | null;
    delete: boolean;
  };
  templates: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  sshKeys: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  roles: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  backups: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  securityGroups: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  cloudProviderAccounts: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  dockerRegistries: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  gitProviders: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  team: {
    create: boolean;
    update: boolean;
    read: boolean;
    delete: boolean;
  };
  type?: ('engineering' | 'management' | 'marketing' | 'finance' | 'sales') | null;
  createdBy?: (string | null) | User;
  tags?: string[] | null;
  isAdminRole?: boolean | null;
  updatedAt: string;
  createdAt: string;
  deletedAt?: string | null;
}